app/requirements.txt
flask==2.3.2
pandas==2.0.3
google-cloud-storage==2.10.0
numpy==1.24.3
xgboost==1.7.6
scikit-learn==1.3.0
